Document end-of-blockdev BDA reverses order of MDA and Static Header
If the start-of-blockdev BDA is overwritten, maybe by an accidental dd,
Stratis also writes the BDA to the end of the blockdev. However, since the
MDA is variable-sized (and we need to read the sigblock to know its size)
put the Static Header in the last 8 sectors, and then define the MDA as
occupying the preceding space, length as defined by the MDA size field
in the sig block.

Signed-off-by: Andy Grover <agrover@redhat.com>