designdoc: Add a JSON schema for metadata
To make sure it's legal JSON.
To, possibly, include in an appendix.
Do not include definitions for TBD parts specified in doc.
Use very limited constraints, like size has to be at least 0. Do not
put significant semantic constraints in the schema.
Assume that UUIDs will be specified in compressed format, without dashes.

Signed-off-by: mulhern <amulhern@redhat.com>