Merges previous site into a static site generator
We no longer need the docs/{guide,faq,relnotes}/Makefile because `zola` takes care of building now

Moved the relnotes, faq, and howto markdown files to locations `zola` is expecting.

Moved the favicon to the static folder so `zola` ignores it.

Sets up the zola scaffolding and theme

adds the built site to the gitiginore

adds instructions on how to build and edit the site

sets up the makefile to handle zola

also adds tests for the zola lyx binaries

removes unnecessary demo files from theme dir

Adds a new target which allows locally serving the website for testing and dev purposes