Fix DBus API Reference documentation regarding AddDevs method
Section 4.2 of the DBus API Reference document indicates a method
'AddDevs' that seems to have been replaced by two new methods:
'AddCacheDevs' and 'AddDataDevs' (or rather, AddDevs became
'AddDataDevs', and a new method 'AddCacheDevs' was added).
This change was introduced via commit
3f55d8c34b1a0df50bfbfb00ded85ce4859b8ad5 but the docs didn't seem to
reflect it. This patch is an attempt to reflect the change.

Finally, I've removed the 'force' option from 'CreatePool',
and 'AddDataDevs'.

Signed-off-by: Jose Castillo <jcastillo@redhat.com>