Add release notes for stratis-cli 1.1.0
Introduce a consistent naming scheme and an _index.html file to try to
organize the pages so that they are displayed in order of most recent.
It turns out that zola has some difficulty interpreting dates, so use
weights instead. Fix up some dates that were wrong, anyway, while here.

Signed-off-by: mulhern <amulhern@redhat.com>