Update D-Bus API reference documentation for v2.0
The stratisd D-Bus API has undergone some significant changes in this version.
This commit updates the documentation in a few key areas: removing
references to the API being unstable, updating the properties API
reference with the correct signatures, fully documenting the new
FetchProperties interface, and adding a section for API versioning.

Co-Authored-By: John Baublitz <jbaublitz@redhat.com>