Update the style guide
* New modification date
* New LyX version
* Do not discuss rustfmt. It has become standard in cargo, and is part of
our CI.
* Changed the third category of imports to "dependencies owned by the
Stratis project". With libcrytpsetup-rs, there are now more than one.
* Leave out discussion of ordering of imports. rustfmt (or isort if Python)
handles all that.
* Indicate that we use merged imports in Rust, although that is obvious.
We wouldn't have to bother if rustfmt enforced, but that enforcement is
not yet on stable.
* Don't claim that using expect() on the output from the Linux kernel
is good behavior. We've proved that it is not.
* Get rid of TBD's.
* Do not refer to PEP8 in Python. We do whatever black and isort enforce
and all that is documented elsewhere.
* Do not make a statement about the minimum supported Python version.
That doesn't really belong in a style guide.

Signed-off-by: mulhern <amulhern@redhat.com>