Revise docs paying special attention to mention of encryption
Remove footnote "in a future version" from "Encryption" item.
Remove brittle "Layer <num>:" numbering in all section headings.
Move encryption specification into its own "Encryption" section and
put a reference to that section in the "Blockdev" layer section.
Get mention of the "Cache Tier" out of the list of layers in the data tier
and put any useful text in the exiting "Cache Tier" section.
Make the "Cache Tier Metadata" section a subsection of the "Cache Tier"
section.
Delete statements about per-filesystem encryption.
Add information about the cipher and MEK size to the set of specifications
in the encryption section.
Get rid of the word "currently" in the description of the CLI.
Fix a typo: RHEL7 -> RHEL8.

Signed-off-by: mulhern <amulhern@redhat.com>