Remove the FAQ document from our website
It's the document that we most ignore. Between the design doc, the D-Bus
reference, the how-to, and the katacoda tutorial we have enough documents
to keep up-to-date.

Also, FAQs are from the olden times, when they were actually lists of
[q]uestions that the maintainers of a project had noticed that people [a]sked
[f]requently. These maintainers chose to make a list of those questions,
and then berate their questioners if they asked a question that was already
answered in the FAQ. Do we ever even check the FAQ to see if the question
has already been asked? Never. It's time for it to go.

Signed-off-by: mulhern <amulhern@redhat.com>