Add author information to blog entries
We simply add the author and organization as italicized text.

The alternatives would require a bunch of extra work:
1. To add author as part of a taxonomy prevents building the website at
all because the hyde theme doesn't supply the template which zola seems to
require for making the author index. We would have to find a different theme
or supply the hyde theme with a template.
2. To add author as an extra would require us to update the version of the
hyde theme we use ourselves to make the extra useful. We would obviously
rather have hyde work on that.

I don't think any of these approaches are worth the trouble.

Overwrite the date where present. The date gets filled in in the expanded
view, and so the duplicate date looks a bit goofy.